requirements.txt
google-adk
google-genai
google-cloud-storage
google-cloud-aiplatform[evaluation]
pandas
Pillow
